+++ src/main/java/co/com/asset/repository/CategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import co.com.asset.model.entity.CategoryEntity;
 
-public interface CategoryRepository extends CrudRepository<CategoryEntity, Integer> {
+public interface CategoryRepository extends CrudRepository<CategoryEntity, Long> {
 
 }
